+++ packages/browser-wallet/src/popup/pages/VerifiableCredential/VerifiableCredentialHooks.tsx
@@ -1,5 +1,5 @@
 import { grpcClientAtom } from '@popup/store/settings';
-import { VerifiableCredential, VerifiableCredentialStatus, VerifiableCredentialSchema } from '@shared/storage/types';
+import { VerifiableCredential, VerifiableCredentialStatus } from '@shared/storage/types';
 import {
     CredentialQueryResponse,
     IssuerMetadata,
@@ -11,6 +11,7 @@ import {
     getCredentialRegistryMetadata,
     getVerifiableCredentialEntry,
     getVerifiableCredentialStatus,
+    VerifiableCredentialSchemaWithFallback,
 } from '@shared/utils/verifiable-credential-helpers';
 import { useAtomValue } from 'jotai';
 import { useEffect, useState } from 'react';
@@ -22,7 +23,7 @@ import {
 import { AsyncWrapper } from '@popup/store/utils';
 import { ConcordiumGRPCClient } from '@concordium/web-sdk';
 import { useTranslation } from 'react-i18next';
-import { logError } from '@shared/utils/log-helpers';
+import { logError, logWarningMessage } from '@shared/utils/log-helpers';
 
 /**
  * Retrieve the on-chain credential status for a verifiable credential in a CIS-4 credential registry contract.
@@ -52,19 +53,31 @@ export function useCredentialStatus(credential: VerifiableCredential) {
  * @throws if no schema is found in storage for the provided credential
  * @returns the credential's schema used for rendering the credential
  */
-export function useCredentialSchema(credential?: VerifiableCredential) {
-    const [schema, setSchema] = useState<VerifiableCredentialSchema>();
+export function useCredentialSchema(
+    credential?: VerifiableCredential
+): VerifiableCredentialSchemaWithFallback | undefined {
+    const [schema, setSchema] = useState<VerifiableCredentialSchemaWithFallback>();
     const schemas = useAtomValue(storedVerifiableCredentialSchemasAtom);
+    const client = useAtomValue(grpcClientAtom);
 
     useEffect(() => {
         if (!schemas.loading && credential) {
-            const schemaValue = schemas.value[credential.credentialSchema.id];
-            if (!schemaValue) {
-                throw new Error(`Attempted to find schema for credentialId: ${credential.id} but none was found!`);
-            }
-            setSchema(schemaValue);
+            const registryContractAddress = getCredentialRegistryContractAddress(credential.id);
+            getCredentialRegistryMetadata(client, registryContractAddress)
+                .then((registryMetadata) => {
+                    let usingFallback = false;
+                    let schemaValue = schemas.value[registryMetadata.credentialSchema.schema.url];
+                    if (!schemaValue) {
+                        // Use the schema we got when originally adding the credential as a fallback for the
+                        // credential, if we do not have the new schema saved yet.
+                        usingFallback = true;
+                        schemaValue = schemas.value[credential.credentialSchema.id];
+                    }
+                    setSchema({ ...schemaValue, usingFallback });
+                })
+                .catch(logError);
         }
-    }, [credential?.id, schemas.loading]);
+    }, [credential?.id, schemas.loading, JSON.stringify(schemas.value)]);
 
     return schema;
 }
@@ -110,24 +123,36 @@ export function useCredentialMetadata(credential?: VerifiableCredential) {
         if (storedMetadata.loading) {
             return;
         }
-        let url;
+
+        let url: string | undefined;
         if (credentialEntry) {
             url = credentialEntry.credentialInfo.metadataUrl.url;
         } else if (!tempMetadata.loading && credential) {
             url = tempMetadata.value[credential.id];
         }
-        if (!url) {
+        if (url === undefined || credential === undefined) {
             return;
         }
+
         const storedCredentialMetadata = storedMetadata.value[url];
-        if (!storedCredentialMetadata) {
+        if (storedCredentialMetadata) {
+            setMetadata(storedCredentialMetadata);
+            return;
+        }
+
+        // The URL we got does not have a corresponding entry in our local storage.
+        // In this case we fallback to using the known "good" value so that we can
+        // still get metadata for this credential.
+        const fallbackCredentialMetadata = storedMetadata.value[credential.metadataUrl];
+        if (!fallbackCredentialMetadata) {
             throw new Error(
-                `Attempted to find credential metadata for credentialId: ${
-                    credentialEntry?.credentialInfo.credentialHolderId || credential?.id
-                } but none was found!`
+                `Attempted to find credential metadata for credentialId: ${credential.id} at URL ${credential.metadataUrl} but none was found!`
             );
         }
-        setMetadata(storedCredentialMetadata);
+        logWarningMessage(
+            `Using fallback credential metadata for credential ${credential.id}. The credential entry metadata URL is [${credentialEntry?.credentialInfo.metadataUrl.url}]`
+        );
+        setMetadata(fallbackCredentialMetadata);
     }, [storedMetadata.loading, tempMetadata.loading, credentialEntry, credential?.id]);
 
     return metadata;