+++ packages/browser-wallet/src/popup/pages/VerifiableCredentialBackup/i18n/en.ts
@@ -2,5 +2,6 @@ export default {
     title: 'Import Web3 ID Credentials',
     noImported: 'No Web3 ID Credentials were imported',
     error: 'Unable to import the chosen file. The file must be a backup created with the same seed phrase.',
+    importButton: 'Select file to import',
     close: 'Close',
 };