+++ packages/browser-wallet/src/shared/storage/types.ts
@@ -289,8 +289,11 @@ export interface VerifiableCredential extends APIVerifiableCredential {
     // Secrets
     signature: string;
     randomness: Record<string, string>;
-    /** index used to derive keys for credential */
+    // Index used to derive keys for credential
     index: number;
+    // The original metadataUrl received when first adding the credential
+    // TODO: The URL should be updated when there are valid updates to the metadata.
+    metadataUrl: string;
 }
 
 interface CredentialSchemaProperty {
@@ -324,16 +327,15 @@ type IdDetails = {
 };
 
 type CredentialSchemaAttributes = {
+    title?: string;
+    description?: string;
+    type: 'object';
     properties: Record<
         string,
         CredentialSchemaProperty | TimestampProperty | SimplePropertyDetails | TimestampPropertyDetails
     >;
     required: string[];
-    title?: string;
-    type: 'object';
-    description?: string;
-    format?: string;
-} & CredentialSchemaProperty;
+};
 
 export interface CredentialSchemaSubject {
     type: string;