+++ packages/browser-wallet/src/shared/utils/verifiable-credential-helpers.ts
@@ -1,4 +1,11 @@
-import { CcdAmount, UpdateContractPayload, ConcordiumGRPCClient, ContractAddress, sha256 } from '@concordium/web-sdk';
+import {
+    CcdAmount,
+    UpdateContractPayload,
+    ConcordiumGRPCClient,
+    ContractAddress,
+    sha256,
+    ConcordiumHdWallet,
+} from '@concordium/web-sdk';
 import * as ed from '@noble/ed25519';
 import {
     MetadataUrl,
@@ -12,6 +19,8 @@ import jsonschema from 'jsonschema';
 import { applyExecutionNRGBuffer, getContractName } from './contract-helpers';
 import { getNet } from './network-helpers';
 import { logError } from './log-helpers';
+
+export type VerifiableCredentialSchemaWithFallback = VerifiableCredentialSchema & { usingFallback: boolean };
 
 /**
  * Extracts the credential holder id from a verifiable credential id (did).
@@ -42,7 +51,7 @@ export function getPublicKeyfromPublicKeyIdentifierDID(did: string) {
 }
 
 /**
- * Extracts the credential registry contract addres from a verifiable credential id (did).
+ * Extracts the credential registry contract address from a verifiable credential id (did).
  * @param credentialId the did for a credential
  * @returns the contract address of the issuing contract of the provided credential id
  */
@@ -233,6 +242,31 @@ export async function getRevokeTransactionExecutionEnergyEstimate(
 }
 
 /**
+ * Determines whether a verifiable credential with the provided key already
+ * exists in the provided contract.
+ *
+ * This is done by invoking the credentialStatus method and checking that the
+ * call succeeded.
+ */
+export async function isVerifiableCredentialInContract(
+    client: ConcordiumGRPCClient,
+    credentialHolderId: string,
+    issuer: ContractAddress
+) {
+    const instanceInfo = await client.getInstanceInfo(issuer);
+    if (instanceInfo === undefined) {
+        throw new Error('Given contract address was not a created instance');
+    }
+    const result = await client.invokeContract({
+        contract: issuer,
+        method: `${getContractName(instanceInfo)}.credentialStatus`,
+        parameter: Buffer.from(credentialHolderId, 'hex'),
+    });
+
+    return result.tag === 'success';
+}
+
+/**
  * Get the status of a verifiable credential in a CIS-4 contract.
  * @param client the GRPC client for accessing a node
  * @param credentialId the id for a verifiable credential
@@ -377,9 +411,6 @@ const verifiableCredentialSchemaSchema = {
                                         description: {
                                             type: 'string',
                                         },
-                                        format: {
-                                            type: 'string',
-                                        },
                                         properties: {
                                             additionalProperties: {
                                                 anyOf: [
@@ -419,10 +450,11 @@ const verifiableCredentialSchemaSchema = {
                                             type: 'string',
                                         },
                                         type: {
+                                            const: 'object',
                                             type: 'string',
                                         },
                                     },
-                                    required: ['description', 'properties', 'required', 'title', 'type'],
+                                    required: ['type', 'properties', 'required'],
                                     type: 'object',
                                 },
                                 id: {
@@ -1133,4 +1165,33 @@ export function getCredentialIdFromSubjectDID(did: string) {
         throw new Error(`Given DID did not follow expected format: ${did}`);
     }
     return split[split.length - 1];
+}
+
+/**
+ * Finds the next unused verifiable credential index, i.e. the next index that has not
+ * been used as a credential holder id in the provided issuer contract.
+ * @param client the GRPC client for accessing a node
+ * @param localIndex the currently best guess on the next index, based on what is currently stored locally
+ * @param issuer the issuer credential registry contract
+ * @param hdWallet the key derivation wallet
+ * @returns the next unused verifiable credential index
+ */
+export async function findNextUnusedVerifiableCredentialIndex(
+    client: ConcordiumGRPCClient,
+    localIndex: number,
+    issuer: ContractAddress,
+    hdWallet: ConcordiumHdWallet
+): Promise<number> {
+    let index = localIndex;
+    let credentialAlreadyExists = true;
+
+    do {
+        const credentialHolderId = hdWallet.getVerifiableCredentialPublicKey(issuer, index).toString('hex');
+        credentialAlreadyExists = await isVerifiableCredentialInContract(client, credentialHolderId, issuer);
+        if (credentialAlreadyExists) {
+            index += 1;
+        }
+    } while (credentialAlreadyExists);
+
+    return index;
 }