+++ packages/browser-wallet-api-helpers/CHANGELOG.md
@@ -1,10 +1,12 @@
 # Changelog
 
-## Unreleased
+## 2.6.0
 
 ### Added
 
 -   `requestAccounts` entrypoint to connect to a list of accounts.
+-   `addWeb3IdCredential` entrypoint to add verifiable credentials.
+-   `requestVerifiablePresentation`entrypoint to prove statements about identities and verifiable credentials.
 
 ## 2.5.1