+++ packages/browser-wallet/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.1.4
+
+### Added
+
+-   The wallet now validates verifiable credential attributes based on their type. String attributes can at most be 31 bytes (UTF-8), integer attributes must fit in a u64 and Date attributes must be between -262144-01-01T00:00:00 and +262143-12-31T23:59:59.999999999Z'.
+
+### Changed
+
+-   Adjusted the schema validation for credential schemas to no longer require title and description. The type is now required to be 'object'.
+
+### Fixed
+
+-   An issue where changing the credential metadata URL to an invalid URL, or a URL that does not contain a credential metadata file, would result in an empty screen.
+-   Issues with a contract switching to an invalid schema or switching the schema to a new URL.
+-   The wallet now ensures that the verifiable credential index used when adding a credential has not already been used in the contract.
+-   An issue where an invalid Date would result in the epoch timestamp instead of returning an error.
+-   Enabled ID statement checks for Web3 ID proof requests containing account credential statements.
+
 ## 1.1.3
 
 ### Changed
@@ -41,6 +59,12 @@
 ### Added
 
 -   Web3IdCredentials support.
+
+## 1.0.7
+
+### Added
+
+-   Support for protocol version 6.
 
 ## 1.0.6